se agrego ventana producto y se modificaron sus propiedades y modelo

+++ TractopartesDeskApp/Models/ProductoModel.cs
@@ -8,13 +8,14 @@ namespace TractopartesDeskApp.Models
 {
     public class ProductoModel
     {
+
         public int p_idproducto { get; set; }
         public string p_productonombre { get; set; } = "";
         public string p_codigopieza { get; set; } = "";
         public string p_descripcion { get; set; } = "";
         public string p_ImagenURL { get; set; } = "";
-        public  float p_precioventa { get; set; }
-        public float p_preciocompra { get; set; }
+        public decimal p_precioventa { get; set; }
+        public decimal  p_preciocompra { get; set; } 
         public int p_cantidad { get; set; }
         public ProveedorModel p_proveedor { get; set; } = new();
         public CategoriaModel p_categoria { get; set; } = new();
